fix: add requireAuth and rateLimiter to AI routes
